+++ src/pacx/clients/__init__.py
@@ -1,9 +1,11 @@
 from .connectors import ConnectorsClient as ConnectorsClient
 from .dataverse import DataverseClient as DataverseClient
 from .power_platform import PowerPlatformClient as PowerPlatformClient
+from .tenant_settings import TenantSettingsClient as TenantSettingsClient
 
 __all__ = [
     "ConnectorsClient",
     "DataverseClient",
     "PowerPlatformClient",
+    "TenantSettingsClient",
 ]